Added: TitaniumDNS.szu
sources.list contains a list of URL for TitaniumDNS Zones Updater each URL provide a list of known domains that are either ads, malicious, dangerous and others.
This file is still work in progress and its content may changes without warnings.

Extension is .szu for Sources Zones Updater, but the file is a simple text file.
Comment a line with the symbol "#" at the beginning of it. No inline comment.